+++ src/FileLocation.ts
@@ -1,0 +1,5 @@
+export interface FileLocation {
+  relativePath: string
+  beginLine?: number
+  endLine?: number
+}